+++ src/CBCMAC.ts
@@ -6,20 +6,22 @@ import {BlockCipher} from "./lib/algorithm/cipher/BlockCipher";
 
 export type CBCMACProps = {
   Cipher: typeof BlockCipher;
-  tagLength?: number;
 };
 
 export function CBCMAC(
-  message: Word32Array|string,
+  plainText: Word32Array|string,
+  associatedData: Word32Array|string,
   key: Word32Array|string,
   iv: Word32Array|null,
+  tagLength?: number,
   props?: Partial<CBCMACProps>,
 ){
   const Cipher = (props && props.Cipher) ? props.Cipher : AES;
   const K = typeof key === "string" ? Utf8.parse(key) : key;
   const N = iv ? iv : new Word32Array([0, 0]);
-  const A = typeof message === "string" ? Utf8.parse(message) : message;
-  const t = props && props.tagLength || 16;
+  const A = typeof associatedData === "string" ? Utf8.parse(associatedData) : associatedData;
+  const P = typeof plainText === "string" ? Utf8.parse(plainText) : plainText;
+  const t = tagLength || 16;
   
-  return CCM.mac(Cipher, K, N, A, undefined, t);
+  return CCM.mac(Cipher, K, N, A, P, t);
 }