+++ src/GMAC.ts
@@ -6,20 +6,20 @@ import {BlockCipher} from "./lib/algorithm/cipher/BlockCipher";
 
 export type GMACProps = {
   Cipher: typeof BlockCipher;
-  tagLength?: number;
 };
 
 export function GMAC(
   message: Word32Array|string,
   key: Word32Array|string,
   iv?: Word32Array,
+  tagLength?: number,
   props?: Partial<GMACProps>,
 ){
   const aad = typeof message === "string" ? Utf8.parse(message) : message;
   const initializingVector = iv ? iv : new Word32Array([0, 0, 0, 0]);
   const Cipher = (props && props.Cipher) ? props.Cipher : AES;
   const wKey = typeof key === "string" ? Utf8.parse(key) : key;
-  const t = props && props.tagLength || 16;
+  const t = tagLength || 16;
   
   return GCM.mac(Cipher, wKey, initializingVector, aad, undefined, t);
 }